requirements.txt
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
jdatetime==3.6.2
pytz==2020.4
requests==2.25.0
